model/requirements.txt
pandas==1.0.5
numpy==1.19.0
PyYAML==5.3.1
scikit_learn==0.23.2
Flask==1.1.1
